Making Visual Awareness
This edit declared gave some logic to how the game will be working. This is only a basic automated version which will be changed.
The low poly drum kit now monitors all its children in an array and chooses one child to have a material that has Emissions enabled so it glows and it easy to see visually.

You can change this by clicking the object that has a different material applied compared to the rest or waiting 3 seconds and it will automatically change.

The timer effect was added so players lose points of they're not fast enough to click the highlighted GameObject.
Note this will be added later on in development.

The next step will be having a basic UI system to show players getting points.

The script was made so the whole team can gather an idea of what we will be working towards and making a Human made array changer that'll be good for different level designs.