Level Unlock System
This Script allows players to unlock different levels. For the time being players do not need to gain anything for levels are going to start to be made soon. When these levels are made we can use a score to define the percentage of each level being done.
This script can be refined later on in development