Big Change
After analysing the current game state i saw that the script the team was using had no room for a ui that the player can play along to.

So now there is a new controller that uses tags and booleans in a game manager to allow players to play drum notes when raycast hits the objects collider.

Players can now earn points when a note is played successfully. This will add the replayability feature for if players don't get the right points when playing a song they cant progress to the next level.

I have made a backup of the old version of the game just in case any ideation or changes suit that version more then the current update.

We also now have a Scan Target however two things need to be fixed.

UI does not show when a build of the game is made and sometimes you can play notes without them being in a trigger area.

The main thing of focus now is to fix these bugs then make songs the player can play along to with their drums and start to build a presentation video