Editing the main Menu and Game States
In this edit i have made the main menu have images on the buttons and we have access to a Options screen where we can change the Quality of the game and the volume of the music.

I also edited how players leave a level after entering it in the level select. This is now done by calculating the Game Managers score int so GameObjects turn off and on outside the level as well as so other levels unlock as im using an If statement to unlock the next level if relevant points have been required.

Players exit the level when the Game can no longer find a GameObject with the tag Music_Notes.

The next thing to do is to have an exit level button as well as a Go to main menu button in the level select